Print the egglog program before the last schedule for multi-pass schedule in `--run-mode egglog`

+++ src/main.rs
@@ -1,7 +1,7 @@
 use clap::Parser;
 use dag_in_context::{EggccConfig, Schedule};
 use eggcc::util::{visualize, InterpMode, LLVMOptLevel, Run, RunMode, TestProgram};
-use std::{ffi::OsStr, path::PathBuf};
+use std::{ffi::OsStr, i64, path::PathBuf};
 
 #[derive(Debug, Parser)]
 struct Args {
@@ -49,8 +49,16 @@ struct Args {
     /// For the eggcc schedule, choose between the sequential and parallel schedules.
     #[clap(long)]
     eggcc_schedule: Option<Schedule>,
+    /// Eggcc by default performs several passes.
+    /// This argument specifies how many passes to run (all passes by default).
+    /// If stop_after_n_passes is negative,
+    /// run [0 ... schedule.len() + stop_after_n_passes] passes.
+    ///
+    /// This flag also works with `--run-mode egglog` mode,
+    /// where it prints the egglog program being processed by the last pass
+    /// this flag specifies.
     #[clap(long)]
-    stop_after_n_passes: Option<usize>,
+    stop_after_n_passes: Option<i64>,
 
     /// Turn off enforcement that the output program uses
     /// memory linearly. This can give an idea of what
@@ -105,7 +113,7 @@ fn main() {
         add_timing: args.add_timing,
         eggcc_config: EggccConfig {
             schedule: args.eggcc_schedule.unwrap_or(Schedule::default()),
-            stop_after_n_passes: args.stop_after_n_passes.unwrap_or(usize::MAX),
+            stop_after_n_passes: args.stop_after_n_passes.unwrap_or(i64::MAX),
             linearity: !args.no_linearity,
         },
     };
